components/MapLayout.js, Splash,js - use buttons instead of links
When appropriate. Tipped off by this eslint warning - should only use
links for links, and not like buttons which invoke javascript code.

> warning  The href attribute is required for an anchor to be keyboard accessible. Provide a
valid, navigable address as the href value. If you cannot provide an href, but still need the element
to resemble a link, use a button and change it with appropriate styles. Learn more:
https://github.com/evcohen/eslint-plugin-jsx-a11y/blob/master/docs/rules/anchor-is-valid.md

Requires re-jigging the styles, to make the default button style look
right, and the use of a .closer class to avoid naming the element
explicitly.