src/*, content/* - rewire so that we display new forestry schema correctly
The field names have changed, and we need to select the way we display
depending on the fields present.

youtube urls need to be displayed if present, else the hero image,
else a blank.

regions need to shown if geojson present, else the map pin, else the
article must be skipped (but warn about this in the console).