styles/templates/article.module.scss work around a chrome bug/feature
The mobile single-column view of articles doesn't scroll in
chrome (mobiles, also mobile emulation).

Briefly: the desktop and default element sequence is text, then,
image, left to right.

So the problem on mobile is to get image then text when displaying in
vertical, single column mode.

Reversing the flexbox ordering with `column-reverse` and aligning it
at the beginning again using `justify-contents: flex-end` works but
not on chrome.

So we now use `flex-direction: column` and set the `.left` and
`.right` elements into order 2, 1

This seems to work on chrome and FF.