gatsby-node.js - attempt to make graphql query work reliably
This may not be working reliably, unfortunately. It was failing
initially, but subsequently seemed to work. I worry that the order of
processing and content of articles can throw the graphql inference off
and make the build fail.

This is retained to document my problems with graphql queries and the
custom_icon field.

If this starts to cause problems I may have to find a more hacky workaround.